automation/requirements.txt
brave-search
openai>=1.0.0
python-dotenv
pyyaml
requests
Pillow
aiohttp
tenacity
